Add and improve OOP
