+++ src/main/java/Parser.java
@@ -1,0 +1,8 @@
+public class Parser {
+
+    public CommandHandler returnCommand(String input) {
+        String instructions = input.split(" ")[0];
+        return new CommandHandler(Commands.fromString(instructions));
+    }
+
+}
